// SqGoods/ClientApp/src/api-client.generated.ts
export interface FilterBoolModel {
}

export interface FilterAndModel extends FilterBoolModel {
    items: FilterBoolModel[];
    type: string;
}

export interface FilterOrModel extends FilterBoolModel {
    items: FilterBoolModel[];
    type: string;
}

export interface FilterPredicateModel extends FilterBoolModel {
    attributeId: string;
}

export interface FilterSelectPredicateModel extends FilterPredicateModel {
    selectValueId: string;
    type: string;
}

export interface FilterSetPredicateModel extends FilterPredicateModel {
    selectValueIds: string[];
    type: string;
}

export interface FilterBoolPredicateModel extends FilterPredicateModel {
    value: boolean;
    type: string;
}

export interface FilterIntBetweenPredicateModel extends FilterPredicateModel {
    from?: number | null;
    to?: number | null;
    type: string;
}